Struct tokio_fs::file::OpenOptions

pub struct OpenOptions(_);

Options and flags which can be used to configure how a file is opened.

This is a specialized version of std::fs::OpenOptions for usage from the Tokio runtime.

From<std::fs::OpenOptions> is implemented for more advanced configuration than the methods provided here.

Methods

impl OpenOptions

pub fn new() -> OpenOptions

Creates a blank new set of options ready for configuration.

All options are initially set to false.

Examples

use tokio::fs::OpenOptions;

let mut options = OpenOptions::new();
let future = options.read(true).open("foo.txt");

pub fn read(&mut self, read: bool) -> &mut OpenOptions

See the underlying read call for details.

pub fn write(&mut self, write: bool) -> &mut OpenOptions

See the underlying write call for details.

pub fn append(&mut self, append: bool) -> &mut OpenOptions

See the underlying append call for details.

pub fn truncate(&mut self, truncate: bool) -> &mut OpenOptions

See the underlying truncate call for details.

pub fn create(&mut self, create: bool) -> &mut OpenOptions

See the underlying create call for details.

pub fn create_new(&mut self, create_new: bool) -> &mut OpenOptions

See the underlying create_new call for details.

pub fn open<P>(&self, path: P) -> OpenFuture<P> where
P: AsRef<Path> + Send + 'static,

Opens a file at path with the options specified by self.

Errors

OpenOptionsFuture results in an error if called from outside of the Tokio runtime or if the underlying open call results in an error.