Struct tokio_fs::Stderr

pub struct Stderr { /* fields omitted */ }

A handle to the standard error stream of a process.

The handle implements the AsyncWrite trait, but beware that concurrent writes to Stderr must be executed with care.

Created by the stderr function.