Crate tokio_fs

Asynchronous filesystem manipulation operations (and stdin, stdout, stderr).

This module contains basic methods and types for manipulating the contents of the local filesystem from within the context of the Tokio runtime.

Tasks running on the Tokio runtime are expected to be asynchronous, i.e., they will not block the thread of execution. Filesystem operations do not satisfy this requirement. In order to perform filesystem operations asynchronously, this library uses the blocking annotation to signal to the runtime that a blocking operation is being performed. This allows the runtime to compensate.

Re-exports

pub use file::File;

Modules

file
Types for working with File.

Structs

OpenOptions
Options and flags which can be used to configure how a file is opened.

Stderr
A handle to the standard error stream of a process.

Stdin
A handle to the standard input stream of a process.

Stdout
A handle to the standard output stream of a process.

Functions

stderr
Constructs a new handle to the standard error of the current process.

stdin
Constructs a new handle to the standard input of the current process.

stdout
Constructs a new handle to the standard output of the current process.